# Add ConvTranspose + BatchNorm fuse pass