@grygielski 2019-10-04
@luotao1 This PR should boost Face model performance. Could you check it?