@grygielski 2019-10-09
I will create tests to satisfy CI_Coverage and post benchmarks of face model today or tomorrow.