@grygielski 2019-10-09
Profiling report:

Before:
```
fps: 50.1784
Event                        Calls       Total       Min.        Max.        Ave.        Ratio.      
thread0::conv2d_transpose    40004       133601      0.939064    17.029      3.3397      0.676819    
thread0::conv2d              40004       36564.1     0.321496    9.88917     0.914012    0.185232    
thread0::tanh                10001       17082.5     1.66925     3.70333     1.70808     0.0865391   
thread0::batch_norm          30003       6029.46     0.089963    1.26321     0.200962    0.030545    
thread0::relu                30003       2309.87     0.031114    0.245071    0.076988    0.0117017   
thread0::fetch               10001       1765.62     0.155819    0.470094    0.176544    0.00894456  
thread0::feed                10001       28.1434     0.001657    0.018939    0.00281405  0.000142573 
thread0::load_combine        1           14.8332     14.8332     14.8332     14.8332     7.51443e-05 
```
After:
```
fps: 52.2322
Event                        Calls       Total       Min.        Max.        Ave.        Ratio.      
thread0::conv2d_transpose    40004       133319      0.929042    21.6668     3.33263     0.702697    
thread0::conv2d              40004       36446.3     0.308973    8.99549     0.911066    0.192101    
thread0::tanh                10001       17035.7     1.67124     3.44101     1.7034      0.0897922   
thread0::fetch               10001       1772.01     0.155506    1.93444     0.177183    0.00933992  
thread0::relu                30003       1109.12     0.020214    0.254751    0.0369671   0.00584598  
thread0::feed                10001       24.5144     0.001554    0.018804    0.00245119  0.000129211 
thread0::load_combine        1           17.8336     17.8336     17.8336     17.8336     9.39975e-05
```
There is around 3% speedup since batch_norm (which execution time ratio is ~3%) is simply deleted from graph without adding additional computing complexity.