@luotao1 2019-10-12
- before: 48.0023ms
```
I1012 18:41:47.405616 23697 helper.h:326] ====== batch latency: 48.0023ms, number of samples: 1, sample latency: 48.0023ms, fps: 20.8323, data type: float ======
W1012 18:41:47.406805 23697 profiler.cc:89] CUDA CUPTI is not enabled

------------------------->     Profiling Report     <-------------------------

Place: CPU
Time unit: ms
Sorted by total time in descending order in the same thread

Event                        Calls       Total       Min.        Max.        Ave.        Ratio.
thread0::conv2d_transpose    400         3795.53     2.88749     17.0744     9.48883     0.794735
thread0::conv2d              400         870.807     0.741943    3.52667     2.17702     0.182336
thread0::batch_norm          300         47.2326     0.109399    0.223445    0.157442    0.00988988
thread0::relu                300         39.0913     0.066213    0.218874    0.130304    0.00818521
thread0::tanh                100         11.7407     0.11043     0.127522    0.117407    0.00245834
thread0::fetch               100         10.9741     0.100498    0.135213    0.109741    0.00229782
thread0::feed                100         0.470948    0.003866    0.011128    0.00470948  9.86103e-05
```
- after: 47.5211ms 
```
I1012 18:53:52.037951 15407 helper.h:326] ====== batch latency: 47.5211ms, number of samples: 1, sample latency: 47.5211ms, fps: 21.0433, data type: float ======
W1012 18:53:52.039032 15407 profiler.cc:89] CUDA CUPTI is not enabled

------------------------->     Profiling Report     <-------------------------

Place: CPU
Time unit: ms
Sorted by total time in descending order in the same thread

Event                        Calls       Total       Min.        Max.        Ave.        Ratio.
thread0::conv2d_transpose    400         3808.75     2.8491      17.7729     9.52186     0.805744
thread0::conv2d              400         874.453     0.738043    3.6237      2.18613     0.184991
thread0::relu                300         21.9281     0.043042    0.146464    0.0730936   0.00463891
thread0::tanh                100         11.532      0.108986    0.148024    0.11532     0.0024396
thread0::fetch               100         9.89461     0.086688    0.121521    0.0989461   0.00209321
thread0::feed                100         0.440194    0.003368    0.013041    0.00440194  9.31235e-05
```